refactor: relocate retry logic
