+++ portal-bridge/src/pandaops.rs
@@ -1,11 +1,16 @@
-use crate::bridge::Retry;
-use crate::constants::{BASE_CL_ENDPOINT, BASE_EL_ENDPOINT};
-use crate::{PANDAOPS_CLIENT_ID, PANDAOPS_CLIENT_SECRET};
 use anyhow::anyhow;
-use ethportal_api::types::jsonrpc::request::JsonRequest;
 use futures::future::join_all;
 use serde_json::{json, Value};
-use tracing::warn;
+use surf::{
+    middleware::{Middleware, Next},
+    Body, Client, Request, Response,
+};
+use tokio::time::{sleep, Duration};
+use tracing::{info, warn};
+
+use crate::constants::{BASE_CL_ENDPOINT, BASE_EL_ENDPOINT};
+use crate::{PANDAOPS_CLIENT_ID, PANDAOPS_CLIENT_SECRET};
+use ethportal_api::types::jsonrpc::request::JsonRequest;
 
 /// Limit the number of requests in a single batch to avoid exceeding the
 /// provider's batch size limit configuration of 100.
@@ -93,5 +98,53 @@ impl PandaOpsMiddleware {
             .recv_string()
             .await;
         result.map_err(|err| anyhow!("Unable to request consensus block from pandaops: {err:?}"))
+    }
+}
+
+#[derive(Debug)]
+pub struct Retry {
+    attempts: u8,
+}
+
+impl Retry {
+    pub fn new(attempts: u8) -> Self {
+        Retry { attempts }
+    }
+}
+
+#[async_trait::async_trait]
+impl Middleware for Retry {
+    async fn handle(
+        &self,
+        mut req: Request,
+        client: Client,
+        next: Next<'_>,
+    ) -> Result<Response, surf::Error> {
+        let mut retry_count: u8 = 0;
+        let body = req.take_body().into_bytes().await?;
+        while retry_count < self.attempts {
+            if retry_count > 0 {
+                info!("Retrying request");
+            }
+            let mut new_req = req.clone();
+            new_req.set_body(Body::from_bytes(body.clone()));
+            if let Ok(val) = next.run(new_req, client.clone()).await {
+                if val.status().is_success() {
+                    return Ok(val);
+                }
+            };
+            retry_count += 1;
+            sleep(Duration::from_millis(100)).await;
+        }
+        Err(surf::Error::from_str(
+            500,
+            "Unable to fetch batch after 3 retries",
+        ))
+    }
+}
+
+impl Default for Retry {
+    fn default() -> Self {
+        Self { attempts: 3 }
     }
 }
